refactor: Improve structure of the app with better organization
