+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"><!--app-html--></div>
-    <script type="module" src="/src/entry-client.jsx"></script>
+    <script type="module" src="/src/client/app.jsx"></script>
   </body>
 </html>
