+++ package.json
@@ -15,7 +15,7 @@
     "migration:undo": "db-migrate --config migrations/db-migrate.json down",
     "build:server": "babel ./src --out-dir build",
     "build:client": "vite build",
-    "build:ssr": "vite build --ssr src/entry-server.jsx --outDir build/server",
+    "build:ssr": "vite build --ssr src/server/app.jsx --outDir build/server",
     "build": "yarn build:client && yarn build:ssr && yarn build:server",
     "clean": "rimraf build && rimraf build-static",
     "start": "NODE_ENV=production node ./build/server/index.js",
