+++ src/server/router.jsx
@@ -3,11 +3,11 @@ import { Route, Routes } from 'react-router-dom';
 
 // Direct imports for SSR (no lazy loading)
 // React.lazy() doesn't work with renderToString() - it requires streaming SSR
-import HomePage from './components/HomePage';
-import ReduxExamplePage from './components/ReduxExamplePage';
-import PostsPage from './components/PostsPage';
-import PostDetailPage from './components/PostDetailPage';
-import NotFound from './components/NotFound';
+import HomePage from '../common/components/HomePage';
+import ReduxExamplePage from '../common/components/ReduxExamplePage';
+import PostsPage from '../common/components/PostsPage';
+import PostDetailPage from '../common/components/PostDetailPage';
+import NotFound from '../common/components/NotFound';
 
 export default function Router() {
   return (
