+++ vitest.config.mjs
@@ -48,8 +48,8 @@ export default defineConfig({
         '**/*.test.jsx',
         '**/index.js',
         '**/*.stories.js',
-        'src/entry-client.jsx',
-        'src/entry-server.jsx',
+        'src/client/app.jsx',
+        'src/server/app.jsx',
         'src/client/index.js',
         'src/server/index.js',
         'babel.config.js',
